# openapi.yaml
openapi: 3.0.3
info:
  title: 唐宋电竞陪玩报单平台API
  description: 唐宋电竞陪玩报单平台后端API文档
  version: 1.0.0
  contact:
    name: 唐宋电竞技术团队
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: https://api.tangsong-esports.com/api/v1
    description: 生产环境
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: https://api.tangsong-esports.com/api/v1
    description: 生产环境
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "操作成功"
        data:
          type: object
...
    Customer:
      type: object
      properties:
        customer_id:
          type: integer
          example: 1
        account:
          type: string
          example: "customer001"
        customer_name:
          type: string
          example: "测试客户"
        contact_method:
          type: string
          example: "微信: test_wx"
        phone_number:
          type: string
          example: "13900139000"
        member_birthday:
          type: string
          format: date-time
          nullable: true
          example: "1990-01-01T00:00:00Z"
        room_code:
          type: string
          example: "ROOM001"
        additional_info1:
          type: string
        additional_info2:
          type: string
        additional_info3:
          type: string
        notes:
          type: string
          example: "VIP客户"
        status:
          type: string
          enum: ["正常", "禁用", "过期"]
          example: "正常"
        is_email_verified:
          type: boolean
          example: false
        is_phone_verified:
          type: boolean
          example: false
        last_login_at:
          type: string
          format: date-time
          nullable: true
        last_login_ip:
          type: string
          example: "127.0.0.1"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        financial_info:
          $ref: '#/components/schemas/CustomerFinancialInfo'
        preferences:
          $ref: '#/components/schemas/CustomerPreferences'
    CustomerFinancialInfo:
      type: object
      properties:
        customer_id:
          type: integer
        initial_real_charge:
          type: number
          format: decimal
          example: 1000.00
        total_consumption:
          type: number
          format: decimal
          example: 500.00
        total_real_charge:
          type: number
          format: decimal
          example: 1500.00
        current_balance:
          type: number
          format: decimal
          example: 1000.00
    CustomerPreferences:
      type: object
      properties:
        customer_id:
          type: integer
        exclusive_discount_type:
          type: string
          example: "固定折扣"
        platform_boss:
          type: string
          example: "平台老板A"
        exclusive_cs:
          type: string
          example: "客服小王"
...
    StandardResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "操作成功"

    # 客户注册登录相关
    CustomerRegisterRequest:
      type: object
      required:
        - account
        - password
        - customer_name
      properties:
        account:
          type: string
          example: "testuser001"
        password:
          type: string
          minLength: 6
          example: "123456"
        customer_name:
          type: string
          example: "测试用户"
        contact_method:
          type: string
          example: "微信: test_wx"
        phone_number:
          type: string
          example: "13900139000"
        member_birthday:
          type: string
          format: date-time
          nullable: true
          example: "1990-01-01T00:00:00Z"
        room_code:
          type: string
          example: "ROOM001"
        additional_info1:
          type: string
        additional_info2:
          type: string
        additional_info3:
          type: string
        notes:
          type: string
    CustomerLoginRequest:
      type: object
      required:
        - account
        - password
      properties:
        account:
          type: string
          example: "testuser001"
        password:
          type: string
          example: "123456"
    CustomerLoginResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expires_in:
          type: integer
          example: 86400
        user:
          $ref: '#/components/schemas/Customer'
    CustomerRefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    CustomerRefreshTokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expires_in:
          type: integer
          example: 86400
        user:
          $ref: '#/components/schemas/Customer'
    CustomerSetPasswordRequest:
      type: object
      required:
        - account
        - password
      properties:
        account:
          type: string
          example: "customer001"
        password:
          type: string
          minLength: 6
          example: "123456"
    CustomerResetPasswordRequest:
      type: object
      required:
        - old_password
        - new_password
      properties:
        old_password:
          type: string
          example: "oldpass123"
        new_password:
          type: string
          minLength: 6
          example: "newpass123"
...
paths:
...
  /customer/register:
    post:
      tags:
        - 客户认证
      summary: 客户注册
      description: 客户自助注册接口
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRegisterRequest'
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CustomerLoginResponse'
        '400':
          description: 参数错误或账号已存在
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      code:
                        type: integer
                        example: 500
                      message:
                        type: string
                        example: "账号已存在"
  /customer/login:
    post:
      tags:
        - 客户认证
      summary: 客户登录
      description: 客户登录接口
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CustomerLoginResponse'
        '400':
          description: 登录失败
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      code:
                        type: integer
                        example: 500
                      message:
                        type: string
                        example: "账号或密码错误"
  /customer/refresh:
    post:
      tags:
        - 客户认证
      summary: 刷新客户访问令牌
      description: 使用刷新令牌获取新的访问令牌
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRefreshTokenRequest'
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CustomerRefreshTokenResponse'
        '401':
          description: 刷新令牌无效或已过期
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      code:
                        type: integer
                        example: 401
                      message:
                        type: string
                        example: "刷新令牌无效或已过期"
  /customer/set-password:
    post:
      tags:
        - 客户认证
      summary: 客户设置密码
      description: 客户设置登录密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerSetPasswordRequest'
      responses:
        '200':
          description: 密码设置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        '400':
          description: 参数错误或账户不存在
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      message:
                        type: string
                        example: "账户不存在"
        '500':
          description: 服务器错误
  /customer/reset-password:
    post:
      tags:
        - 客户认证
      summary: 客户重置密码
      description: 客户重置登录密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerResetPasswordRequest'
      responses:
        '200':
          description: 密码重置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        '400':
          description: 参数错误或账户不存在
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      message:
                        type: string
                        example: "账户不存在"
        '500':
          description: 服务器错误
...
  /customer/profile:
    get:
      tags:
        - 客户认证
      summary: 获取客户个人信息
      description: 获取当前登录客户的个人信息
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Customer'

    put:
      tags:
        - 客户认证
      summary: 更新客户个人信息
      description: 更新当前登录客户的个人信息
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRegisterRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Customer'
...